dataset/classification/requirements.txt
Pillow>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
matplotlib>=3.5.0